api_fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
cryptography
pandas
openpyxl
python-docx
reportlab
jinja2
requests
python-multipart
passlib[bcrypt]
bcrypt==3.2.2
PyJWT